requirements.txt
spotipy
python-dotenv
streamlit
pandas